Safari-609.1.11.1/LayoutTests/crypto/workers/subtle/ec-import-spki-key-expected.txt

[Worker] Test importing a SPKI ECDH key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-import-spki-key.js
[Worker] Importing a key...
PASS [Worker] publicKey.toString() is '[object CryptoKey]'
PASS [Worker] publicKey.type is 'public'
PASS [Worker] publicKey.extractable is true
PASS [Worker] publicKey.algorithm.name is 'ECDH'
PASS [Worker] publicKey.algorithm.namedCurve is 'P-256'
PASS [Worker] publicKey.usages is [ ]
PASS successfullyParsed is true

TEST COMPLETE
